+++ native/tests/test_cpp_header.cpp
@@ -4,6 +4,13 @@
 static_assert(noexcept(ams_mel_get_abi_version(nullptr)));
 static_assert(std::is_standard_layout_v<ams_mel_abi_version_v1>);
 static_assert(std::is_standard_layout_v<ams_mel_ir_return_result_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_u32_span_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_string_view_span_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_ir_scan_type_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_ir_scan_param_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_ir_mode_command_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_ir_bit_command_v1>);
+static_assert(std::is_standard_layout_v<ams_mel_ir_config_set_command_v1>);
 static_assert(noexcept(ams_mel_ir_return_request_close(nullptr, nullptr, 0, nullptr)));
 
 int main()